# Conflicts:
#	core/parachain/validator/backing_implicit_view.cpp
#	core/parachain/validator/fragment_tree.hpp
#	core/parachain/validator/impl/parachain_processor.cpp
#	core/parachain/validator/parachain_processor.hpp
#	core/parachain/validator/prospective_parachains.hpp
#	core/parachain/validator/prospective_parachains/constraints.cpp

+++ core/consensus/babe/impl/threshold_util.cpp
@@ -16,7 +16,8 @@ namespace kagome::consensus::babe {
                                AuthorityIndex authority_index) {
     BOOST_ASSERT(authority_index < authorities.size());
 
-    double float_point_ratio = double(ratio.first) / ratio.second;
+    double float_point_ratio =
+        static_cast<double>(ratio.first) / static_cast<double>(ratio.second);
 
     using boost::adaptors::transformed;
     double theta = double(authorities[authority_index].weight)
